chore: rename app payment to account payment
