+++ bd_api/apps/account/apps.py
@@ -3,8 +3,9 @@ from django.apps import AppConfig
 
 
 class AccountConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
     name = "bd_api.apps.account"
+    verbose_name = "Contas"
+    default_auto_field = "django.db.models.BigAutoField"
 
     def ready(self):
         import bd_api.apps.account.signals  # noqa
