+++ bd_api/apps/account_auth/apps.py
@@ -3,5 +3,6 @@ from django.apps import AppConfig
 
 
 class AuthConfig(AppConfig):
-    default_auto_field = "django.db.models.BigAutoField"
     name = "bd_api.apps.account_auth"
+    verbose_name = "Autenticação e Autorização Interna"
+    default_auto_field = "django.db.models.BigAutoField"
