+++ bd_api/apps/account_payment/apps.py
@@ -3,9 +3,9 @@ from djstripe.apps import DjstripeAppConfig
 
 
 class PaymentConfig(DjstripeAppConfig):
-    verbose_name = "Stripe"
+    verbose_name = "Pagamentos"
 
     def ready(self):
         super().ready()
-        import bd_api.apps.payment.signals  # noqa
-        import bd_api.apps.payment.webhooks  # noqa
+        import bd_api.apps.account_payment.signals  # noqa
+        import bd_api.apps.account_payment.webhooks  # noqa
