+++ bd_api/apps/api/v1/apps.py
@@ -3,7 +3,6 @@ from django.apps import AppConfig
 
 
 class ApiConfig(AppConfig):
-    app_label = "v1"
-    verbose_name = " V1"
     name = "bd_api.apps.api.v1"
+    verbose_name = " API"
     default_auto_field = "django.db.models.BigAutoField"
