+++ bd_api/urls.py
@@ -42,6 +42,6 @@ urlpatterns = [
     path("search/debug/", include("haystack.urls")),
     path("dataset/", DatasetRedirectView.as_view()),
     path("dataset_redirect/", DatasetRedirectView.as_view()),
-    path("payment/", include("bd_api.apps.payment.urls")),
+    path("payment/", include("bd_api.apps.account_payment.urls")),
 ]
 urlpatterns += static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
